Move types to nested mod
